fix build size issues by moving constants to central file
