+++ src/features/session_replay/constants.js
@@ -1,4 +1,4 @@
-import { MODE } from '../../common/session/session-entity'
+import { MODE } from '../../common/session/constants'
 import { FEATURE_NAMES } from '../../loaders/features/features'
 
 export const FEATURE_NAME = FEATURE_NAMES.sessionReplay
