+++ tests/assets/rrweb-instrumented.html
@@ -4,42 +4,55 @@
   PDX-License-Identifier: Apache-2.0
 -->
 <html>
-<head>
-  <title>RUM Unit Test</title>
-  <style>
-    .left{
-      position: absolute; left: 50px; top: 200px;
-    }
-    .right {
-      position: absolute; right: 50px; top: 200px;
-    }
-  </style>
-  <link rel="stylesheet" type="text/css" href="style.css">
-  {init}
-  {config}
-  {loader}
-</head>
-<body>
-  this is a page that provides several types of elements with selectors that session_replay can interact with based on how it is configured
-  <hr />
-  <hr />
-  <textarea id="plain"></textarea>
-  <textarea id="ignore" class="nr-ignore"></textarea>
-  <textarea id="block" class="nr-block"></textarea>
-  <textarea id="mask" class="nr-mask"></textarea>
-  <textarea id="nr-block" data-nr-block></textarea>
-  <textarea id="other-block" data-other-block></textarea>
-  <input type="password" id="pass-input">
-  <input type="text" id="text-input">
-  <hr />
-  <button onclick="moveImage()">Click</button>
-  <img src="https://upload.wikimedia.org/wikipedia/commons/d/d7/House_of_Commons_Chamber_1.png" />
-  <a href="./rrweb-instrumented.html" target="_blank">New Tab</a>
-  <script>
-    function moveImage(){
-      document.querySelector("img").classList.toggle("left")
-      document.querySelector("img").classList.toggle("right")
-    }
-  </script>
-</body>
+  <head>
+    <title>RUM Unit Test</title>
+    <style>
+      .left {
+        position: absolute;
+        left: 50px;
+        top: 200px;
+      }
+      .right {
+        position: absolute;
+        right: 50px;
+        top: 200px;
+      }
+    </style>
+    <link rel="stylesheet" type="text/css" href="style.css" />
+    {init} {config} {loader}
+    <script>
+      window.wasPreloaded = false;
+        window.addEventListener("load", () => {
+          try{
+          const replayEvents = Object.values(newrelic.initializedAgents)[0].features.session_replay.recorder.getEvents();
+          window.wasPreloaded = !!replayEvents.events.length && replayEvents.type === "preloaded"
+          } catch(err){
+            // do nothing because it failed to get the recorder events -- which means they dont exist
+          }
+        });
+    </script>
+  </head>
+  <body>
+    this is a page that provides several types of elements with selectors that session_replay can interact with based on how it is configured
+    <hr />
+    <hr />
+    <textarea id="plain"></textarea>
+    <textarea id="ignore" class="nr-ignore"></textarea>
+    <textarea id="block" class="nr-block"></textarea>
+    <textarea id="mask" class="nr-mask"></textarea>
+    <textarea id="nr-block" data-nr-block></textarea>
+    <textarea id="other-block" data-other-block></textarea>
+    <input type="password" id="pass-input" />
+    <input type="text" id="text-input" />
+    <hr />
+    <button onclick="moveImage()">Click</button>
+    <img src="https://upload.wikimedia.org/wikipedia/commons/d/d7/House_of_Commons_Chamber_1.png" />
+    <a href="./rrweb-instrumented.html" target="_blank">New Tab</a>
+    <script>
+      function moveImage() {
+        document.querySelector("img").classList.toggle("left");
+        document.querySelector("img").classList.toggle("right");
+      }
+    </script>
+  </body>
 </html>
